requirements.txt
Django
celery
django-celery-beat
django-celery-results
django-configurations
django-filter
django-redis
django-storages
djangorestframework
djangorestframework-simplejwt
djangorestframework-gis
Markdown
psycopg2-binary
python-dotenv
drf-spectacular
